% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{AccountSettings-class}
\alias{AccountSettings}
\alias{AccountSettings-class}
\title{AccountSettings}
\description{
The return value from \code{\link{change_account_settings}} and
\code{\link{get_account_settings}}.
}
\details{
Each slot length must be \code{0} or \code{1}.
}
\section{Slots}{

\describe{
\item{\code{account}}{Object of class \code{"\link{RippleAddress}"}. The Ripple
address of the account in question.}

\item{\code{regular_key}}{Object of class \code{"\link{RippleAddress}"}. The hash of
an optional additional public key that can be used for signing and
verifying transactions.}

\item{\code{domain}}{Object of class \code{"character"}. The domain associated with
this account. The \code{ripple.txt} file can be looked up to verify this
information.}

\item{\code{email_hash}}{Object of class \code{"\link{Hash128}"}. The MD5 128-bit
hash of the account owner's email address.}

\item{\code{message_key}}{Object of class \code{"character"}. An optional public key,
represented as hex, that can be set to allow others to send encrypted
messages to the account owner.}

\item{\code{transfer_rate}}{Object of class \code{"\link{UINT32}"}.}

\item{\code{require_destination_tag}}{Object of class \code{"logical"}. If set to
\code{TRUE} incoming payments will only be validated if they include a
\code{destination_tag}. This may be used primarily by gateways that operate
exclusively with hosted wallets.}

\item{\code{require_authorization}}{Object of class \code{"logical"}. If set to
\code{TRUE} incoming trustlines will only be validated if this account
first creates a trustline to the counterparty with the authorized flag set
to \code{TRUE}. This may be used by gateways to prevent accounts unknown to
them from holding currencies they issue.}

\item{\code{disallow_xrp}}{Object of class \code{"logical"}. If set to \code{TRUE}
incoming XRP payments will not be allowed.}

\item{\code{password_spent}}{Object of class \code{"logical"}.}

\item{\code{disable_master}}{Object of class \code{"logical"}.}

\item{\code{transaction_sequence}}{Object of class \code{"\link{UINT32}"}. The last
sequence number of a validated transaction created by this account.}

\item{\code{trustline_count}}{Object of class \code{"\link{UINT32}"}. The number of
trustlines owned by this account. This value does not include incoming
trustlines where this account has not explicitly reciprocated trust.}

\item{\code{ledger}}{Object of class \code{"numeric"}. The index number of the ledger
containing these account settings or, in the case of historical queries, of
the transaction that modified these settings.}

\item{\code{hash}}{Object of class \code{"\link{Hash256}"}. If this object was
  returned by a historical query this value will be the hash of the
  transaction that modified these settings. The transaction hash is used
  throughout the Ripple Protocol to uniquely identify a particular
  transaction.}
}}

