% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{as.character,Balance-method}
\alias{as.character,Balance-method}
\title{as.character for Balance class}
\usage{
\S4method{as.character}{Balance}(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
as.character for Balance class
}

