% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{as.character-method}
\alias{as.character,Amount-method}
\alias{as.character,Balance-method}
\alias{as.character-method}
\title{Coerce to character}
\usage{
\S4method{as.character}{Balance}(x, ...)

\S4method{as.character}{Amount}(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Coerce an object to \code{"character"} class.
}

