% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{as.numeric,Amount-method}
\alias{as.numeric,Amount-method}
\title{as.numeric for Amount class}
\usage{
\S4method{as.numeric}{Amount}(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extracts the slot \code{"value"} from an \code{\link{Amount}} object
}

