% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{as.numeric,Balance-method}
\alias{as.numeric,Balance-method}
\title{as.numeric for Balance class}
\usage{
\S4method{as.numeric}{Balance}(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extracts the slot \code{"value"} from a \code{\link{Balance}} object
}

