% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{c,Balance-method}
\alias{c,Balance-method}
\title{Combine for Balance class}
\usage{
\S4method{c}{Balance}(x, ..., recursive = FALSE)
}
\arguments{
\item{x}{object to be concatenated.}

\item{recursive}{ignored}

\item{...}{objects to be concatenated.}
}
\description{
Combine for Balance class
}

