% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{c-method}
\alias{c,Balance-method}
\alias{c-method}
\title{Combine values}
\usage{
\S4method{c}{Balance}(x, ..., recursive = FALSE)
}
\arguments{
\item{x}{object to be concatenated.}

\item{...}{objects to be concatenated.}

\item{recursive}{ignored}
}
\description{
Combine arguments with the same class.
}

