% Generated by roxygen2 (4.0.0): do not edit by hand
\name{change_account_settings}
\alias{change_account_settings}
\title{Updating Account Settings}
\usage{
change_account_settings(address, secret, transfer_rate, domain, message_key,
  email_hash, disallow_xrp = NA, require_authorization = NA,
  require_destination_tag = NA, password_spent = NA)
}
\arguments{
\item{address}{The Ripple address of the desired account.}

\item{secret}{The secret key for your Ripple account.}

\item{transfer_rate}{The rate charged each time a holder of currency issued
by this account transfers some funds. The default and minimum rate is
"1.0"; a rate of "1.01" is a 1\% charge on top of the amount being
transferred. Up to nine decimal places are supported.}

\item{domain}{The domain name associated with this account.}

\item{message_key}{An optional public key, represented as a hex string, that
can be used to allow others to send encrypted messages to the account
owner.}

\item{email_hash}{The MD5 128-bit hash of the account owner's email address,
if known.}

\item{disallow_xrp}{If this is set to \code{TRUE}, payments in XRP will not
be allowed.}

\item{require_authorization}{If this is set to \code{TRUE}, incoming
trustlines will only be validated if this account first creates a trustline
to the counterparty with the authorized flag set to \code{TRUE}. This may
be used by gateways to prevent accounts unknown to them from holding
currencies they issue.}

\item{require_destination_tag}{If this is set to \code{TRUE}, incoming
payments will only be validated if they include a \code{destination_tag}
value. Note that this is used primarily by gateways that operate
exclusively with hosted wallets.}

\item{password_spent}{\code{TRUE} if the password has been "spent", else
  \code{FALSE}.}
}
\value{
An object of class \code{"\link{AccountSettings}"}
}
\description{
Change an account's settings
}

