% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check_payment_status}
\alias{check_payment_status}
\title{Confirming a Payment}
\usage{
check_payment_status(status_url, address, client_resource_id, hash)
}
\arguments{
\item{status_url}{Return value from \code{\link{submit_payment}}.}

\item{address}{The Ripple address for the source account. Ignored if
\code{status_url} is provided.}

\item{client_resource_id}{Provided to \code{\link{submit_payment}}. Ignored
if \code{status_url} is provided.}

\item{hash}{The transaction hash for the desired payment. Ignored if
  \code{status_url} or \code{client_resource_id} is provided.}
}
\value{
An object of class \code{"\link{Payment}"}
}
\description{
To confirm that your payment has been submitted successfully, you can use
this function.
}

