% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{Currency-class}
\alias{Currency}
\alias{Currency-class}
\title{Currency}
\description{
The three-character code or hex string used to denote currencies
}
\examples{
USD <- Currency("USD")
XAU <- Currency("015841551A748AD2C1F76FF6ECB0CCCD00000000")
}

