% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{currency}
\alias{currency}
\alias{currency,Amount-method}
\alias{currency,Balance-method}
\alias{currency,Trustline-method}
\alias{currency<-}
\alias{currency<-,Amount,Currency-method}
\alias{currency<-,Amount,character-method}
\alias{currency<-,Balance,Currency-method}
\alias{currency<-,Balance,character-method}
\alias{currency<-,Trustline,Currency-method}
\alias{currency<-,Trustline,character-method}
\title{Get/set \code{currency} value}
\usage{
currency(object)

\S4method{currency}{Balance}(object)

\S4method{currency}{Amount}(object)

\S4method{currency}{Trustline}(object)

\S4method{currency}{Balance,Currency}(object) <- value

\S4method{currency}{Amount,Currency}(object) <- value

\S4method{currency}{Trustline,Currency}(object) <- value

\S4method{currency}{Balance,character}(object) <- value

\S4method{currency}{Amount,character}(object) <- value

\S4method{currency}{Trustline,character}(object) <- value
}
\arguments{
\item{object}{Object with currency slot.}

\item{value}{Object of class \code{"\link{Currency}"} or class
  \code{"character"}.}
}
\value{
Object of class \code{"\link{Currency}"}.
}
\description{
Access the \code{currency} slot.
}
\examples{
x <- Amount(1, "USD"); x
currency(x) <- "EUR"; x
}

