% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{destination_balance_changes}
\alias{destination_balance_changes}
\alias{destination_balance_changes,Payment-method}
\title{Get destination balance changes}
\usage{
destination_balance_changes(object)

\S4method{destination_balance_changes}{Payment}(object)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}
}
\value{
Object of class \code{"\link{Amount}"}.
}
\description{
Access the \code{destination_balance_changes} slot.
}

