% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{destination_tag}
\alias{destination_tag}
\alias{destination_tag,Payment-method}
\alias{destination_tag<-}
\alias{destination_tag<-,Payment,ANY-method}
\alias{destination_tag<-,Payment,UINT32-method}
\title{Get/set \code{destination_tag} value}
\usage{
destination_tag(object)

\S4method{destination_tag}{Payment}(object)

\S4method{destination_tag}{Payment,UINT32}(object) <- value

\S4method{destination_tag}{Payment,ANY}(object) <- value
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{value}{Object of class \code{"\link{UINT32}"} or class
  \code{"ANY"}.}
}
\value{
Object of class \code{"\link{UINT32}"}.
}
\description{
Access the \code{destination_tag} slot.
}

