% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{extract-method}
\alias{[,Amount-method}
\alias{[,Balance-method}
\alias{[,Payment-method}
\alias{[,Trustline-method}
\alias{[<-,Balance,ANY,missing,Balance-method}
\alias{[<-,Trustline,ANY,missing,Trustline-method}
\alias{extract-method}
\title{Extract/replace parts of object}
\usage{
\S4method{[}{Balance}(x, i)

\S4method{[}{Trustline}(x, i)

\S4method{[}{Amount}(x, i)

\S4method{[}{Payment}(x, i)

\S4method{[}{Balance,ANY,missing,Balance}(x, i) <- value

\S4method{[}{Trustline,ANY,missing,Trustline}(x, i) <- value
}
\arguments{
\item{x}{object from which to extract element(s) or in which to replace
element(s).}

\item{i}{indices specifying elements to extract or replace.}

\item{value}{typically an array-like \R object of a similar class as
  \code{x}.}
}
\description{
Extract or replace parts of an object.
}

