% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate_account}
\alias{generate_account}
\title{Generating Accounts}
\usage{
generate_account()
}
\value{
A list with the address and the secret for a potential new account
}
\description{
There are two steps to making a new account on the Ripple network: randomly
creating the keys for that account, and sending it enough XRP to meet the
account reserve. Generating the keys can be done offline, since it does not
affect the network at all. To make it easy, this function can generate
account keys for you.
}
\details{
The generated account does not exist in the ledger until it receives enough
XRP to meet the account reserve.
}

