% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generate_uuid}
\alias{generate_uuid}
\title{Create Client Resource ID}
\usage{
generate_uuid()
}
\value{
An object of class \code{"\link{ResourceId}"}
}
\description{
This endpoint creates a universally unique identifier (UUID) value which can
be used to calculate a client resource ID for a payment. This can be useful
if the application does not have a UUID generator handy.
}

