% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_account_balances}
\alias{get_account_balances}
\title{Account Balances}
\usage{
get_account_balances(address, currency, counterparty)
}
\arguments{
\item{address}{The Ripple address of the desired account}

\item{currency}{Three letter currency denominations}

\item{counterparty}{The Ripple address of the counterparty trusted}
}
\value{
An object of class \code{"\link{Balance}"}
}
\description{
Retrieve the current balances for the given Ripple account.
}
\details{
The parameters \code{currency} and \code{counterparty} are supported to
provide additional filtering.
}
\examples{
root_account <- RippleAddress("rHb9CJAWyB4rj91VRWn96DkukG4bwdtyTh")
\dontrun{
get_account_balances(root_account)}
}

