% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_account_payments}
\alias{get_account_payments}
\title{Payment History}
\usage{
get_account_payments(address, source_account, destination_account,
  exclude_failed = FALSE, direction, earliest_first = FALSE, start_ledger,
  end_ledger, results_per_page = 10, page = 1)
}
\arguments{
\item{address}{The Ripple address of the desired account}

\item{source_account}{Filter the results to only include payments sent by the
given account.}

\item{destination_account}{Filter the results to only include payments
received by the given account.}

\item{exclude_failed}{If set to \code{TRUE}, the results will only include
payments which were successfully validated and written into the ledger.
Otherwise, failed payments will be included. Defaults to \code{FALSE}.}

\item{direction}{Limit the results to only include the given type of
payments. The following direction values are currently supported:
\code{incoming}, \code{outgoing}, \code{pending}}

\item{earliest_first}{If set to \code{TRUE}, the payments will be returned in
ascending date order. Otherwise, the payments will be returned in
descending date order (ie, the most recent payment will be returned first).
Defaults to \code{FALSE}.}

\item{start_ledger}{The index for the starting ledger. If
\code{earliest_first} is \code{TRUE}, this will be the oldest ledger to be
queried; otherwise, it will be the most recent ledger. Defaults to the
first ledger in the \code{rippled} server's database.}

\item{end_ledger}{The index for the ending ledger. If \code{earliest_first}
is \code{TRUE}, this will be the most recent ledger to be queried;
otherwise, it will be the oldest ledger. Defaults to the most recent ledger
in the \code{rippled} server's database.}

\item{results_per_page}{The maximum number of payments to be returned at
once. Defaults to 10.}

\item{page}{The page number to be returned. The first page of results will
  have page number 1, the second page will have page number 2, and so on.
  Defaults to 1.}
}
\value{
A list of objects of class \code{"\link{Payment}"}
}
\description{
This API endpoint can be used to browse through an account's payment history
and also used to confirm specific payments after a payment has been
submitted.
}
\details{
If the server returns fewer than \code{results_per_page} payments, then there
are no more pages of results to be returned. Otherwise, increment the page
number and re-issue the query to get the next page of results.

Note that the \code{ripple-rest} API has to retrieve the full list of
payments from the server and then filter them before returning them back to
the caller. This means that there is no speed advantage to specifying more
filter values.
}

