% Generated by roxygen2 (4.0.0): do not edit by hand
\name{get_account_settings}
\alias{get_account_settings}
\title{Account Settings}
\usage{
get_account_settings(address)
}
\arguments{
\item{address}{The Ripple address of the desired account}
}
\value{
An object of class \code{"\link{AccountSettings}"}
}
\description{
You can retrieve an account's settings. The server will return a list of the
current settings in force for the given account.
}

