% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_account_trustlines}
\alias{get_account_trustlines}
\title{Reviewing Trustlines}
\usage{
get_account_trustlines(address, currency, counterparty)
}
\arguments{
\item{address}{The Ripple address of the desired account}

\item{currency}{Three letter currency denominations}

\item{counterparty}{The Ripple address of the counterparty trusted}
}
\value{
An object of class \code{"\link{Trustline}"}
}
\description{
Retrieves all trustlines associated with the Ripple address.
}
\details{
The parameters \code{currency} and \code{counterparty} are supported to
provide additional filtering.
}

