% Generated by roxygen2 (4.0.0): do not edit by hand
\name{get_notification}
\alias{get_notification}
\title{Checking Notifications}
\usage{
get_notification(address, hash)
}
\arguments{
\item{address}{The Ripple address of the desired account}

\item{hash}{Transaction hash}
}
\value{
An object of class \code{"\link{Notification}"}
}
\description{
This endpoint will grab the notification based on the specific transaction
hash specified. Once called the notification object retrieved will provide
information on the type of transaction and also the previous and next
notifications will be shown as well. The \code{previous_notification_url} and
\code{next_notification_url} can be used to walk up and down the notification
queue. Once the \code{next_notification_url} is empty that means you have the
most current notification, this applies for the
\code{previous_notification_url} similarly when it's empty as it means you
are holding the earliest notification available on the rippled you are
connecting to.
}

