% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_server_status}
\alias{get_server_status}
\title{Get Server Status}
\usage{
get_server_status()
}
\value{
A list of lists
}
\description{
Retrieve information about the current status of the \code{ripple-rest} API
and the \code{rippled} server it is connected to.
}

