% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_transaction}
\alias{get_transaction}
\title{Retrieve Ripple Transaction}
\usage{
get_transaction(hash)
}
\arguments{
\item{hash}{Transaction hash}
}
\value{
A list
}
\description{
While the Ripple-REST API is a high-level API built on top of the
\code{rippled} server, there are times when you may need to access an
underlying Ripple transaction rather than dealing with the Ripple-REST data
format. When you need to do this, you can retrieve a transaction with this
function.
}

