% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getCurrency,Amount-method}
\alias{getCurrency,Amount-method}
\title{getCurrency for Amount class}
\usage{
\S4method{getCurrency}{Amount}(object)
}
\arguments{
\item{object}{Object of class \code{"\link{Amount}"}}
}
\description{
Extracts the \code{currency} slot.
}

