% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{getCurrency,Balance-method}
\alias{getCurrency,Balance-method}
\title{getCurrency for Balance class}
\usage{
\S4method{getCurrency}{Balance}(object)
}
\arguments{
\item{object}{Object of class \code{"\link{Balance}"}}
}
\description{
Extracts the \code{currency} slot.
}

