% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getCurrency}
\alias{getCurrency}
\alias{getCurrency,Amount-method}
\alias{getCurrency,Balance-method}
\alias{getCurrency,Trustline-method}
\title{getCurrency}
\usage{
getCurrency(object, ...)

\S4method{getCurrency}{Balance}(object)

\S4method{getCurrency}{Amount}(object)

\S4method{getCurrency}{Trustline}(object)
}
\arguments{
\item{object}{Object with currency slot.}

\item{...}{ignored}
}
\value{
An object of class \code{"\link{Currency}"}.
}
\description{
Extracts the \code{currency} slot.
}

