% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{getCurrency,Trustline-method}
\alias{getCurrency,Trustline-method}
\title{getCurrency for Trustline class}
\usage{
\S4method{getCurrency}{Trustline}(object)
}
\arguments{
\item{object}{Object of class \code{"\link{Trustline}"}}
}
\description{
Extracts the \code{currency} slot.
}

