% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getDestinationBalanceChanges}
\alias{getDestinationBalanceChanges}
\alias{getDestinationBalanceChanges,Payment-method}
\title{getDestinationBalanceChanges}
\usage{
getDestinationBalanceChanges(object, ...)

\S4method{getDestinationBalanceChanges}{Payment}(object)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{...}{ignored}
}
\value{
An object of class \code{"link{Amount}"}.
}
\description{
Extracts the \code{destination_balance_changes} slot.
}

