% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getDestinationBalanceChanges,Payment-method}
\alias{getDestinationBalanceChanges,Payment-method}
\title{getDestinationBalanceChanges for Payment class}
\usage{
\S4method{getDestinationBalanceChanges}{Payment}(object)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}
}
\description{
Extracts the \code{destination_balance_changes} slot.
}

