% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getDestinationBalanceChanges}
\alias{getDestinationBalanceChanges}
\title{getDestinationBalanceChanges}
\usage{
getDestinationBalanceChanges(object, ...)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{...}{ignored}
}
\description{
Extracts the \code{destination_balance_changes} slot.
}

