% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getSourceAmount}
\alias{getSourceAmount}
\alias{getSourceAmount,Payment-method}
\title{getSourceAmount}
\usage{
getSourceAmount(object, ...)

\S4method{getSourceAmount}{Payment}(object)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{...}{ignored}
}
\value{
An object of class \code{"link{Amount}"}.
}
\description{
Extracts the \code{source_amount} slot.
}

