% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getSourceAmount,Payment-method}
\alias{getSourceAmount,Payment-method}
\title{getSourceAmount for Payment class}
\usage{
\S4method{getSourceAmount}{Payment}(object)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}
}
\description{
Extracts the \code{source_amount} slot.
}

