% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getSourceAmount}
\alias{getSourceAmount}
\title{getSourceAmount}
\usage{
getSourceAmount(object, ...)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{...}{ignored}
}
\description{
Extracts the \code{source_amount} slot.
}

