% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getSourceBalanceChanges}
\alias{getSourceBalanceChanges}
\alias{getSourceBalanceChanges,Payment-method}
\title{getSourceBalanceChanges}
\usage{
getSourceBalanceChanges(object, ...)

\S4method{getSourceBalanceChanges}{Payment}(object)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{...}{ignored}
}
\value{
An object of class \code{"\link{Amount}"}.
}
\description{
Extracts the \code{source_balance_changes} slot.
}

