% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{getSourceBalanceChanges,Payment-method}
\alias{getSourceBalanceChanges,Payment-method}
\title{getSourceBalanceChanges for Payment class}
\usage{
\S4method{getSourceBalanceChanges}{Payment}(object)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}
}
\description{
Extracts the \code{source_balance_changes} slot.
}

