% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getSourceBalanceChanges}
\alias{getSourceBalanceChanges}
\title{getSourceBalanceChanges}
\usage{
getSourceBalanceChanges(object, ...)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{...}{ignored}
}
\description{
Extracts the \code{source_balance_changes} slot.
}

