% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{getTransferRate,AccountSettings-method}
\alias{getTransferRate,AccountSettings-method}
\title{getTransferRate for AccountSettings class}
\usage{
\S4method{getTransferRate}{AccountSettings}(object)
}
\arguments{
\item{object}{Object of class \code{"\link{AccountSettings}"}.}
}
\description{
Extracts the \code{transfer_rate} slot.
}

