% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getTransferRate}
\alias{getTransferRate}
\alias{getTransferRate,AccountSettings-method}
\title{getTransferRate}
\usage{
getTransferRate(object, ...)

\S4method{getTransferRate}{AccountSettings}(object)
}
\arguments{
\item{object}{Object of class \code{"\link{AccountSettings}"}.}

\item{...}{ignored}
}
\value{
The transfer rate as \code{"numeric"}.
}
\description{
Extracts the \code{transfer_rate} slot.
}

