% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getTransferRate}
\alias{getTransferRate}
\title{getTransferRate}
\usage{
getTransferRate(object, ...)
}
\arguments{
\item{object}{Object of class \code{"\link{AccountSettings}"}.}

\item{...}{ignored}
}
\description{
Extracts the \code{transfer_rate} slot.
}

