% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{has_ledger}
\alias{has_ledger}
\alias{has_ledger,Payment-method}
\title{Check for valid \code{ledger} slot}
\usage{
has_ledger(object)

\S4method{has_ledger}{Payment}(object)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Checks whether an object has a valid \code{ledger} slot.
}

