% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Hash128-class}
\alias{Hash128}
\alias{Hash128-class}
\title{Hash128 class}
\description{
The hex representation of a 128-bit hash.
}
\details{
A character vector where each element must match the regular expression
\code{"^$|^[A-Fa-f0-9]{32}$"}.
}

