% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Hash256-class}
\alias{Hash256}
\alias{Hash256-class}
\title{Hash256 class}
\description{
The hex representation of a 256-bit hash.
}
\details{
A character vector where each element must match the regular expression
\code{"^$|^[A-Fa-f0-9]{64}$"}.
}

