% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{hasLedger}
\alias{hasLedger}
\alias{hasLedger,Payment-method}
\title{hasLedger}
\usage{
hasLedger(object, ...)

\S4method{hasLedger}{Payment}(object)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{...}{ignored}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Checks whether an object has a valid \code{ledger} slot.
}

