% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{hasLedger,Payment-method}
\alias{hasLedger,Payment-method}
\title{hasLedger for Payment class}
\usage{
\S4method{hasLedger}{Payment}(object)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}
}
\description{
Checks whether an object has a valid \code{ledger} slot.
}

