% Generated by roxygen2 (4.0.0): do not edit by hand
\name{hasLedger}
\alias{hasLedger}
\title{hasLedger}
\usage{
hasLedger(object, ...)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{...}{ignored}
}
\description{
Checks whether an object has a valid \code{ledger} slot.
}

