% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{invoice_id}
\alias{invoice_id}
\alias{invoice_id,Payment-method}
\alias{invoice_id<-}
\alias{invoice_id<-,Payment,Hash256-method}
\alias{invoice_id<-,Payment,character-method}
\title{Get/set invoice id}
\usage{
invoice_id(object)

\S4method{invoice_id}{Payment}(object)

\S4method{invoice_id}{Payment,Hash256}(object) <- value

\S4method{invoice_id}{Payment,character}(object) <- value
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{value}{Object of class \code{"\link{Hash256}"} or class
  \code{"character"}.}
}
\value{
Object of class \code{"\link{Hash256}"}.
}
\description{
Access the \code{invoice_id} slot.
}

