% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is_server_connected}
\alias{is_server_connected}
\title{Check Connection State}
\usage{
is_server_connected()
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Checks to see if the Ripple-REST API is currently connected to a
\code{rippled} server, and is ready to be used. This provides a quick and
easy way to check to see if the API is up and running, before attempting to
process transactions.
}

