% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{length,Amount-method}
\alias{length,Amount-method}
\title{Length for Amount class}
\usage{
\S4method{length}{Amount}(x)
}
\arguments{
\item{x}{an \R object.  For replacement, a vector or factor.}
}
\description{
Length for Amount class
}

