% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{length-method}
\alias{length,Amount-method}
\alias{length,Balance-method}
\alias{length,Notification-method}
\alias{length,Payment-method}
\alias{length,Trustline-method}
\alias{length-method}
\title{Length of object}
\usage{
\S4method{length}{Balance}(x)

\S4method{length}{Notification}(x)

\S4method{length}{Trustline}(x)

\S4method{length}{Amount}(x)

\S4method{length}{Payment}(x)
}
\arguments{
\item{x}{an \R object.  For replacement, a vector or factor.}
}
\description{
Get the length of an object.
}
\examples{
x <- Amount(1, "USD")
length(x)
}

