% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{length,Notification-method}
\alias{length,Notification-method}
\title{Length for Notification class}
\usage{
\S4method{length}{Notification}(x)
}
\arguments{
\item{x}{an \R object.  For replacement, a vector or factor.}
}
\description{
Length for Notification class
}

