% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{length,Payment-method}
\alias{length,Payment-method}
\title{Length for Payment class}
\usage{
\S4method{length}{Payment}(x)
}
\arguments{
\item{x}{an \R object.  For replacement, a vector or factor.}
}
\description{
Length for Payment class
}

