% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{length,Trustline-method}
\alias{length,Trustline-method}
\title{Length for Trustline class}
\usage{
\S4method{length}{Trustline}(x)
}
\arguments{
\item{x}{an \R object.  For replacement, a vector or factor.}
}
\description{
Length for Trustline class
}

