% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{Notification-class}
\alias{Notification}
\alias{Notification-class}
\title{Notification}
\description{
Notification
}
\section{Slots}{

\describe{
\item{\code{account}}{Object of class \code{"\link{RippleAddress}"}. The Ripple
address of the account to which the notification pertains.}

\item{\code{type}}{Object of class \code{"character"}. The resource type this
notification corresponds to. Possible values are "payment", "order",
"trustline", "accountsettings".}

\item{\code{direction}}{Object of class \code{"character"}. The direction of the
transaction, from the perspective of the account being queried. Possible
values are "incoming", "outgoing", and "passthrough".}

\item{\code{state}}{Object of class \code{"character"}. The state of the transaction
from the perspective of the Ripple Ledger. Possible values are "validated"
and "failed".}

\item{\code{result}}{Object of class \code{"character"}. The rippled code indicating
the success or failure type of the transaction. The code "tesSUCCESS"
indicates that the transaction was successfully validated and written into
the Ripple Ledger. All other codes will begin with the following prefixes:
"tec", "tef", "tel", or "tej".}

\item{\code{ledger}}{Object of class \code{"numeric"}. The index number of the ledger
containing the validated or failed transaction. Failed payments will only
be written into the Ripple Ledger if they fail after submission to a
rippled and a Ripple Network fee is claimed.}

\item{\code{hash}}{Object of class \code{"\link{Hash256}"}. The 256-bit hash of the
transaction. This is used throughout the Ripple protocol as the unique
identifier for the transaction.}

\item{\code{timestamp}}{Object of class \code{"POSIXct"}. The timestamp representing
when the transaction was validated and written into the Ripple ledger.}

\item{\code{transaction_url}}{Object of class \code{"character"}. An URL that can be
used to fetch the full resource this notification corresponds to.}

\item{\code{previous_notification_url}}{Object of class \code{"character"}. An URL
that can be used to fetch the notification that preceded this one
chronologically.}

\item{\code{next_notification_url}}{Object of class \code{"character"}. An URL that
  can be used to fetch the notification that followed this one
  chronologically.}
}}

