% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{partial_payment}
\alias{partial_payment}
\alias{partial_payment,Payment-method}
\alias{partial_payment<-}
\alias{partial_payment<-,Payment,logical-method}
\title{Get/set flag for partial payment}
\usage{
partial_payment(object)

\S4method{partial_payment}{Payment}(object)

\S4method{partial_payment}{Payment,logical}(object) <- value
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{value}{Object of class \code{"logical"}.}
}
\value{
Object of class \code{"logical"}.
}
\description{
Access the \code{partial_payment} slot.
}

