% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Payment-class}
\alias{Payment}
\alias{Payment-class}
\title{Payment}
\description{
A flattened Payment object used by the ripple-rest API.
}
\details{
All of the 11 first slot lengths must be equal.

Each of the other slot lengths must be \code{0} or \code{1}.

All elements of \code{source_slippage} must be greater than or equal to
\code{0}.

The slots \code{partial_payment} and \code{no_direct_ripple} cannot contain
\code{NA}.
}
\section{Slots}{

\describe{
\item{\code{source_account}}{Object of class \code{"\link{RippleAddress}"}. The
Ripple account address of the Payment sender.}

\item{\code{source_tag}}{Object of class \code{"\link{UINT32}"}. An unsigned 32-bit
integer most commonly used to refer to a sender's hosted account at a
Ripple gateway.}

\item{\code{source_amount}}{Object of class \code{"\link{Amount}"}. An optional
amount that can be specified to constrain cross-currency payments.}

\item{\code{source_slippage}}{Object of class \code{"numeric"}. An optional cushion
for the \code{source_amount} to increase the likelihood that the payment
will succeed. The \code{source_account} will never be charged more than
\code{source_amount@value} + \code{source_slippage}.}

\item{\code{destination_account}}{Object of class \code{"\link{RippleAddress}"}.}

\item{\code{destination_tag}}{Object of class \code{"\link{UINT32}"}. An unsigned
32-bit integer most commonly used to refer to a receiver's hosted account
at a Ripple gateway.}

\item{\code{destination_amount}}{Object of class \code{"\link{Amount}"}. The amount
the \code{destination_account} will receive.}

\item{\code{invoice_id}}{Object of class \code{"\link{Hash256}"}. A 256-bit hash that
can be used to identify a particular payment.}

\item{\code{paths}}{Object of class \code{"character"}. A "stringified" version of
the Ripple PathSet structure that users should treat as opaque.}

\item{\code{partial_payment}}{Object of class \code{"logical"}. A boolean that, if
set to \code{TRUE}, indicates that this payment should go through even if
the whole amount cannot be delivered because of a lack of liquidity or
funds in the \code{source_account} account.}

\item{\code{no_direct_ripple}}{Object of class \code{"logical"}. A boolean that can
be set to \code{TRUE} if paths are specified and the sender would like the
Ripple Network to disregard any direct paths from the \code{source_account}
to the \code{destination_account}. This may be used to take advantage of an
arbitrage opportunity or by gateways wishing to issue balances from a hot
wallet to a user who has mistakenly set a trustline directly to the hot
wallet.}

\item{\code{direction}}{Object of class \code{"character"}. The direction of the
payment, from the perspective of the account being queried. Possible values
are \code{"incoming"}, \code{"outgoing"}, and \code{"passthrough"}.}

\item{\code{state}}{Object of class \code{"character"}. The state of the payment from
the perspective of the Ripple Ledger. Possible values are
\code{"validated"} and \code{"failed"} and \code{"new"} if the payment has
not been submitted yet.}

\item{\code{result}}{Object of class \code{"character"}. The rippled code indicating
the success or failure type of the payment. The code \code{"tesSUCCESS"}
indicates that the payment was successfully validated and written into the
Ripple Ledger. All other codes will begin with the following prefixes:
\code{"tec"}, \code{"tef"}, \code{"tel"}, or \code{"tej"}.}

\item{\code{ledger}}{Object of class \code{"numeric"}. The index number of the ledger
containing the validated or failed payment. Failed payments will only be
written into the Ripple Ledger if they fail after submission to a rippled
and a Ripple Network fee is claimed.}

\item{\code{hash}}{Object of class \code{"\link{Hash256}"}. The 256-bit hash of the
payment. This is used throughout the Ripple protocol as the unique
identifier for the transaction.}

\item{\code{timestamp}}{Object of class \code{"POSIXct"}. The timestamp representing
when the payment was validated and written into the Ripple ledger.}

\item{\code{fee}}{Object of class \code{"numeric"}. The Ripple Network transaction
fee, represented in whole XRP (NOT "drops", or millionths of an XRP, which
is used elsewhere in the Ripple protocol).}

\item{\code{source_balance_changes}}{Object of class \code{"\link{Amount}"}. Parsed
from the validated transaction metadata, this represents all of the changes
to balances held by the \code{source_account}. Most often this will have
one amount representing the Ripple Network fee and, if the
\code{source_amount} was not XRP, one amount representing the actual
\code{source_amount} that was sent.}

\item{\code{destination_balance_changes}}{Object of class \code{"\link{Amount}"}.
  Parsed from the validated transaction metadata, this represents the changes
  to balances held by the \code{destination_account}. For those receiving
  payments this is important to check because if the \code{partial_payment}
  flag is set this value may be less than the \code{destination_amount}.}
}}

