% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{ResourceId-class}
\alias{ResourceId}
\alias{ResourceId-class}
\title{ResourceId}
\description{
A client-supplied unique identifier (ideally a UUID) for this transaction
used to prevent duplicate payments and help confirm the transaction's final
status. All ASCII printable characters are allowed. Note that 256-bit hex
strings are disallowed because of the potential confusion with transaction
hashes.
}

