% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{ripplerestr}
\alias{ripplerestr}
\alias{ripplerestr-package}
\title{Ripple REST Client for R}
\description{
The \code{ripple-rest} API makes it easy to access the Ripple system via a
RESTful web interface.
}
\details{
The default url is \code{http://localhost:5990/}. You can change it by
setting the option \code{"ripplerestr.url"}.
}
\examples{
options("ripplerestr.url" = "http://example.com/")
}
\references{
\url{https://dev.ripple.com/}
}

