% Generated by roxygen2 (4.0.0): do not edit by hand
\name{set_account_trustline}
\alias{set_account_trustline}
\title{Granting a Trustline}
\usage{
set_account_trustline(address, secret, amount, allows_rippling = NA, limit,
  currency, counterparty)
}
\arguments{
\item{address}{The Ripple address of the desired account}

\item{secret}{The secret key for your Ripple account.}

\item{amount}{Object of class \code{"\link{Amount}"}. The limit, currency and
counterparty for the trustline.}

\item{allows_rippling}{\code{TRUE} or \code{FALSE}. Allow rippling or not for
the trustline.}

\item{limit}{A number indicating the maximum you are willing to trust.
Ignored if \code{amount} is provided.}

\item{currency}{Three letter currency denomination. Ignored if \code{amount}
is provided.}

\item{counterparty}{Ripple address of the counterparty trusted. Ignored if
  \code{amount} is provided.}
}
\value{
An object of class \code{"\link{Trustline}"}
}
\description{
A trustline can also updated and simply set with a currency, amount,
counterparty combination by submitting to this endpoint.
}

