% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setDestinationTag}
\alias{setDestinationTag}
\alias{setDestinationTag,Payment-method}
\title{setDestinationTag}
\usage{
setDestinationTag(object, tag, ...)

\S4method{setDestinationTag}{Payment}(object, tag)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{tag}{Object of class \code{"\link{UINT32}"}.}

\item{...}{ignored}
}
\value{
\code{object} with \code{destination_tag} set to \code{tag}.
}
\description{
Sets the \code{destination_tag} slot.
}

