% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{setDestinationTag,Payment-method}
\alias{setDestinationTag,Payment-method}
\title{setDestinationTag for Payment class}
\usage{
\S4method{setDestinationTag}{Payment}(object, tag)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{tag}{Object of class \code{"\link{UINT32}"}.}
}
\description{
Sets the \code{destination_tag} slot.
}

