% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setDestinationTag}
\alias{setDestinationTag}
\title{setDestinationTag}
\usage{
setDestinationTag(object, tag, ...)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{tag}{Object of class \code{"\link{UINT32}"}.}

\item{...}{ignored}
}
\description{
Sets the \code{destination_tag} slot.
}

