% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setInvoiceId}
\alias{setInvoiceId}
\alias{setInvoiceId,Payment-method}
\title{setInvoiceId}
\usage{
setInvoiceId(object, id, ...)

\S4method{setInvoiceId}{Payment}(object, id)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{id}{Object of class \code{"\link{Hash256}"}.}

\item{...}{ignored}
}
\value{
\code{object} with \code{invoice_id} set to \code{id}.
}
\description{
Sets the \code{invoice_id} slot.
}

