% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setInvoiceId,Payment-method}
\alias{setInvoiceId,Payment-method}
\title{setInvoiceId for Payment class}
\usage{
\S4method{setInvoiceId}{Payment}(object, id)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{id}{Object of class \code{"\link{Hash256}"}.}
}
\description{
Sets the \code{invoice_id} slot.
}

