% Generated by roxygen2 (4.0.0): do not edit by hand
\name{setInvoiceId}
\alias{setInvoiceId}
\title{setInvoiceId}
\usage{
setInvoiceId(object, id, ...)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{id}{Object of class \code{"\link{Hash256}"}.}

\item{...}{ignored}
}
\description{
Sets the \code{invoice_id} slot.
}

