% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setNoDirectRipple}
\alias{setNoDirectRipple}
\alias{setNoDirectRipple,Payment-method}
\title{setNoDirectRipple}
\usage{
setNoDirectRipple(object, x, ...)

\S4method{setNoDirectRipple}{Payment}(object, x)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{x}{Object of class \code{"logical"}.}

\item{...}{ignored}
}
\value{
\code{object} with \code{no_direct_ripple} set to \code{x}.
}
\description{
Sets the \code{no_direct_ripple} slot.
}

