% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setNoDirectRipple,Payment-method}
\alias{setNoDirectRipple,Payment-method}
\title{setNoDirectRipple for Payment class}
\usage{
\S4method{setNoDirectRipple}{Payment}(object, x)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{x}{Object of class \code{"logical"}.}
}
\description{
Sets the \code{no_direct_ripple} slot.
}

