% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setNoDirectRipple}
\alias{setNoDirectRipple}
\title{setNoDirectRipple}
\usage{
setNoDirectRipple(object, x, ...)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{x}{Object of class \code{"logical"}.}

\item{...}{ignored}
}
\description{
Sets the \code{no_direct_ripple} slot.
}

