% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setPartialPayment}
\alias{setPartialPayment}
\alias{setPartialPayment,Payment-method}
\title{setPartialPayment}
\usage{
setPartialPayment(object, x, ...)

\S4method{setPartialPayment}{Payment}(object, x)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{x}{Object of class \code{"logical"}.}

\item{...}{ignored}
}
\value{
\code{object} with \code{partial_payment} set to \code{x}.
}
\description{
Sets the \code{partial_payment} slot.
}

