% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setPartialPayment,Payment-method}
\alias{setPartialPayment,Payment-method}
\title{setPartialPayment for Payment class}
\usage{
\S4method{setPartialPayment}{Payment}(object, x)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{x}{Object of class \code{"logical"}.}
}
\description{
Sets the \code{partial_payment} slot.
}

