% Generated by roxygen2 (4.0.0): do not edit by hand
\name{setPartialPayment}
\alias{setPartialPayment}
\title{setPartialPayment}
\usage{
setPartialPayment(object, x, ...)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{x}{Object of class \code{"logical"}.}

\item{...}{ignored}
}
\description{
Sets the \code{partial_payment} slot.
}

