% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setSourceSlippage}
\alias{setSourceSlippage}
\alias{setSourceSlippage,Payment-method}
\title{setSourceSlippage}
\usage{
setSourceSlippage(object, x, ...)

\S4method{setSourceSlippage}{Payment}(object, x)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{x}{Object of class \code{"numeric"}.}

\item{...}{ignored}
}
\value{
\code{object} with \code{source_slippage} set to \code{x}.
}
\description{
Sets the \code{source_slippage} slot.
}

