% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{setSourceSlippage,Payment-method}
\alias{setSourceSlippage,Payment-method}
\title{setSourceSlippage for Payment class}
\usage{
\S4method{setSourceSlippage}{Payment}(object, x)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{x}{Object of class \code{"numeric"}.}
}
\description{
Sets the \code{source_slippage} slot.
}

