% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setSourceSlippage}
\alias{setSourceSlippage}
\title{setSourceSlippage}
\usage{
setSourceSlippage(object, x, ...)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{x}{Object of class \code{"numeric"}.}

\item{...}{ignored}
}
\description{
Sets the \code{source_slippage} slot.
}

