% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setSourceTag}
\alias{setSourceTag}
\alias{setSourceTag,Payment-method}
\title{setSourceTag}
\usage{
setSourceTag(object, tag, ...)

\S4method{setSourceTag}{Payment}(object, tag)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{tag}{Object of class \code{"\link{UINT32}"}.}

\item{...}{ignored}
}
\value{
\code{object} with \code{source_tag} set to \code{tag}.
}
\description{
Sets the \code{source_tag} slot.
}

