% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setSourceTag,Payment-method}
\alias{setSourceTag,Payment-method}
\title{setSourceTag for Payment class}
\usage{
\S4method{setSourceTag}{Payment}(object, tag)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{tag}{Object of class \code{"\link{UINT32}"}.}
}
\description{
Sets the \code{source_tag} slot.
}

