% Generated by roxygen2 (4.0.0): do not edit by hand
\name{setSourceTag}
\alias{setSourceTag}
\title{setSourceTag}
\usage{
setSourceTag(object, tag, ...)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{tag}{Object of class \code{"\link{UINT32}"}.}

\item{...}{ignored}
}
\description{
Sets the \code{source_tag} slot.
}

