% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show,AccountSettings-method}
\alias{show,AccountSettings-method}
\title{Show for AccountSettings class}
\usage{
\S4method{show}{AccountSettings}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
Show for AccountSettings class
}

