% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show-method}
\alias{show,AccountSettings-method}
\alias{show,Amount-method}
\alias{show,Balance-method}
\alias{show-method}
\title{Show object}
\usage{
\S4method{show}{Balance}(object)

\S4method{show}{Amount}(object)

\S4method{show}{AccountSettings}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
Display an object.
}

