% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{source_amount}
\alias{source_amount}
\alias{source_amount,Payment-method}
\title{Get source amount}
\usage{
source_amount(object)

\S4method{source_amount}{Payment}(object)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}
}
\value{
Object of class \code{"\link{Amount}"}.
}
\description{
Access the \code{source_amount} slot.
}

