% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{source_balance_changes}
\alias{source_balance_changes}
\alias{source_balance_changes,Payment-method}
\title{Get source balance changes}
\usage{
source_balance_changes(object)

\S4method{source_balance_changes}{Payment}(object)
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}
}
\value{
Object of class \code{"\link{Amount}"}.
}
\description{
Access the \code{source_balance_changes} slot.
}

