% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{source_slippage}
\alias{source_slippage}
\alias{source_slippage,Payment-method}
\alias{source_slippage<-}
\alias{source_slippage<-,Payment,numeric-method}
\title{Get/set source slippage}
\usage{
source_slippage(object)

\S4method{source_slippage}{Payment}(object)

\S4method{source_slippage}{Payment,numeric}(object) <- value
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{value}{Object of class \code{"numeric"}.}
}
\value{
Object of class \code{"numeric"}.
}
\description{
Access the \code{source_slippage} slot.
}

