% Generated by roxygen2 (4.0.2): do not edit by hand
\name{submit_payment}
\alias{submit_payment}
\title{Submitting a Payment}
\usage{
submit_payment(payment, secret, client_resource_id)
}
\arguments{
\item{payment}{The \code{"\link{Payment}"} object to be submitted.}

\item{secret}{The secret or private key for your Ripple account.}

\item{client_resource_id}{Will uniquely identify this payment. This is a
  36-character UUID (universally unique identifier) value which will uniquely
  identify this payment within the \code{ripple-rest} API. Note that you can
  use \code{\link{generate_uuid}} to calculate a UUID value if you do not
  have a UUID generator readily available.}
}
\value{
A named list. The first element is the \code{"client_resource_id"}
  you gave. The second element is named \code{"status_url"} and can be used
  with \code{\link{check_payment_status}}.
}
\description{
Before you can submit a payment, you will need to have three pieces of
information: \code{payment}, \code{secret} and \code{client_resource_id}.
}

