% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{transfer_rate}
\alias{transfer_rate}
\alias{transfer_rate,AccountSettings-method}
\title{Get transfer rate}
\usage{
transfer_rate(object)

\S4method{transfer_rate}{AccountSettings}(object)
}
\arguments{
\item{object}{Object of class \code{"\link{AccountSettings}"}.}
}
\value{
The transfer rate as \code{"numeric"}.
}
\description{
Access the \code{transfer_rate} slot.
}

