% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Trustline-class}
\alias{Trustline}
\alias{Trustline-class}
\title{Trustline}
\description{
A simplified Trustline object used by the \code{ripple-rest} API.
}
\details{
All slot lengths must be equal.
}
\section{Slots}{

\describe{
\item{\code{account}}{Object of class \code{"\link{RippleAddress}"}. The account from
whose perspective this trustline is being viewed.}

\item{\code{counterparty}}{Object of class \code{"\link{RippleAddress}"}. The other
party in this trustline.}

\item{\code{currency}}{Object of class \code{"\link{Currency}"}. The code of the
currency in which this trustline denotes trust.}

\item{\code{limit}}{Object of class \code{"numeric"}. The maximum value of the
currency that the account may hold issued by the counterparty.}

\item{\code{reciprocated_limit}}{Object of class \code{"numeric"}. The maximum value
of the currency that the counterparty may hold issued by the account.}

\item{\code{account_allows_rippling}}{Object of class \code{"logical"}. If
\code{TRUE} it indicates that the account allows pairwise rippling out
through this trustline.}

\item{\code{counterparty_allows_rippling}}{Object of class \code{"logical"}. If
\code{TRUE} it indicates that the counterparty allows pairwise rippling out
through this trustline.}

\item{\code{ledger}}{Object of class \code{"numeric"}. The index number of the ledger
containing this trustline or, in the case of historical queries, of the
transaction that modified this Trustline.}

\item{\code{hash}}{Object of class \code{"\link{Hash256}"}. If this object was
  returned by a historical query this value will be the hash of the
  transaction that modified this Trustline. The transaction hash is used
  throughout the Ripple Protocol to uniquely identify a particular
  transaction.}
}}

