% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{as.numeric-method}
\alias{as.numeric,Amount-method}
\alias{as.numeric,Balance-method}
\alias{as.numeric-method}
\title{Coerce to numeric}
\usage{
\S4method{as.numeric}{Balance}(x, ...)

\S4method{as.numeric}{Amount}(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extract the slot \code{"value"} from an object.
}

