% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{Balance-class}
\alias{Balance}
\alias{Balance-class}
\title{Balance}
\description{
A simplified representation of an account Balance
}
\section{Slots}{

\describe{
\item{\code{value}}{Object of class \code{"numeric"}. The quantity of the currency.}

\item{\code{currency}}{Object of class \code{"\link{Currency}"}. The currency
expressed as a three-character code.}

\item{\code{counterparty}}{Object of class \code{"character"}. The Ripple account
  address of the currency's issuer or gateway, or an empty string if the
  currency is XRP.}
}}

