% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Currency-class}
\alias{Currency}
\alias{Currency-class}
\title{Currency class}
\description{
The three-character code or hex string used to denote currencies.
}
\details{
A character vector where each element must match the regular expression
\code{"^([a-zA-Z0-9]{3}|[A-Fa-f0-9]{40})$"}.
}
\examples{
USD <- Currency("USD")
XAU <- Currency("015841551A748AD2C1F76FF6ECB0CCCD00000000")
}

