% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{no_direct_ripple}
\alias{no_direct_ripple}
\alias{no_direct_ripple,Payment-method}
\alias{no_direct_ripple<-}
\alias{no_direct_ripple<-,Payment,logical-method}
\title{Get/set flag for no direct ripple}
\usage{
no_direct_ripple(object)

\S4method{no_direct_ripple}{Payment}(object)

\S4method{no_direct_ripple}{Payment,logical}(object) <- value
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{value}{Object of class \code{"logical"}.}
}
\value{
Object of class \code{"logical"}.
}
\description{
Access the \code{no_direct_ripple} slot.
}

