% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{ResourceId-class}
\alias{ResourceId}
\alias{ResourceId-class}
\title{ResourceId class}
\description{
A client-supplied unique identifier (ideally a UUID) for this transaction
used to prevent duplicate payments and help confirm the transaction's final
status. All ASCII printable characters are allowed. Note that 256-bit hex
strings are disallowed because of the potential confusion with transaction
hashes.
}
\details{
A character vector where each element must match the regular expression
\code{"^(?!$|^[A-Fa-f0-9]{64})[ -~]{1,255}$"}.
}

