% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{ripplerestr-package}
\alias{ripplerestr}
\alias{ripplerestr-package}
\title{Ripple REST Client for R}
\description{
The \href{https://dev.ripple.com/ripple-rest.html}{\code{ripple-rest}} API
makes it easy to access the Ripple system via a RESTful web interface. The \R
package \pkg{ripplerestr} uses the \pkg{\link{httr}} package to communicate
with \code{ripple-rest}.
}
\details{
The default url is \code{http://localhost:5990/}. You can change it by
setting the option \code{"ripplerestr.url"}.
}
\examples{
options("ripplerestr.url" = "http://example.com/")
}

