% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{source_tag}
\alias{source_tag}
\alias{source_tag,Payment-method}
\alias{source_tag<-}
\alias{source_tag<-,Payment,ANY-method}
\alias{source_tag<-,Payment,UINT32-method}
\title{Get/set source tag}
\usage{
source_tag(object)

\S4method{source_tag}{Payment}(object)

\S4method{source_tag}{Payment,UINT32}(object) <- value

\S4method{source_tag}{Payment,ANY}(object) <- value
}
\arguments{
\item{object}{Object of class \code{"\link{Payment}"}.}

\item{value}{Object of class \code{"\link{UINT32}"} or class
  \code{"ANY"}.}
}
\value{
Object of class \code{"\link{UINT32}"}.
}
\description{
Access the \code{source_tag} slot.
}

